+++ src/EditWarningApi.php
@@ -5,19 +5,24 @@ namespace EditWarning;
 use ApiBase;
 use Exception;
 use MediaWiki\MediaWikiServices;
-use User;
 
 class EditWarningApi extends ApiBase {
 
 	public function execute() {
+		$user = $this->getUser();
+		if ( !$user->isRegistered() ) {
+			$this->dieWithError( 'editwarning-error-notloggedin', 'notloggedin' );
+		}
+
 		// Get Params
 		$params = $this->extractRequestParams();
 		$ewAction = $params['ewaction'];
 		$articleID = $params['articleid'];
 		$sectionID = $params['section'];
-		$user = User::newFromName( $params['user'] );
 
-		// Create EditWarning Object of current user and article ID.
+		// Create EditWarning Object of the authenticated user and article ID.
+		// The lock is always attributed to the session user; it must never be taken
+		// from a request parameter, or any caller could act on behalf of anyone else.
 		$ew = new EditWarning();
 		$ew->setUserID( $user->getId() );
 		$ew->setUserName( $user->getName() );
@@ -56,6 +61,27 @@ class EditWarningApi extends ApiBase {
 	}
 
 	/**
+	 * @inheritDoc
+	 */
+	public function isWriteMode() {
+		return true;
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function mustBePosted() {
+		return true;
+	}
+
+	/**
+	 * @inheritDoc
+	 */
+	public function needsToken() {
+		return 'csrf';
+	}
+
+	/**
 	 * @return array allowed parameters
 	 */
 	public function getAllowedParams() {
@@ -75,11 +101,6 @@ class EditWarningApi extends ApiBase {
 				ApiBase::PARAM_REQUIRED => true,
 				ApiBase::PARAM_HELP_MSG => 'apihelp-editwarning-apiparam-ewaction',
 			],
-			'user' => [
-				ApiBase::PARAM_TYPE => 'string',
-				ApiBase::PARAM_REQUIRED => true,
-				ApiBase::PARAM_HELP_MSG => 'apihelp-editwarning-apiparam-user',
-			]
 		];
 	}
 
@@ -88,10 +109,10 @@ class EditWarningApi extends ApiBase {
 	 */
 	protected function getExamplesMessages() {
 		return [
-			'action=editwarning&ewaction=lock&articleid=1&user=Admin' => 'apihelp-editwarning-example-1',
-			'action=editwarning&ewaction=unlock&articleid=56&user=Admin' => 'apihelp-editwarning-example-2',
-			'action=editwarning&ewaction=lock&section=3&articleid=1&user=Admin' => 'apihelp-editwarning-example-3',
-			'action=editwarning&ewaction=unlock&section=3&articleid=56&user=Admin' => 'apihelp-editwarning-example-4',
+			'action=editwarning&ewaction=lock&articleid=1' => 'apihelp-editwarning-example-1',
+			'action=editwarning&ewaction=unlock&articleid=56' => 'apihelp-editwarning-example-2',
+			'action=editwarning&ewaction=lock&section=3&articleid=1' => 'apihelp-editwarning-example-3',
+			'action=editwarning&ewaction=unlock&section=3&articleid=56' => 'apihelp-editwarning-example-4',
 		];
 	}
 }